New encoding of in tuple compatible format.
Timestamp and value are now encoded as message pack
arrays, meaning they can unpack as tuples in Python.

Handling this, is a new graphite encoder having one method.
It's a struct which keeps a reference to the backing messages
pack encoder. The struct was introduced to keep the message
pack encoder around, for performance, and because the message
pack encoder struct does not allow access to the buffer. Thus
making extending difficult.

Move main test to top level.