Fix trimming.
Trimming was not aligned for the data structure put in place
by the recent tuple encoder. Solution is to use a corresponding
decoder.

To allow easier testing, the trimming function was split into
a part handling the Redis interaction and a port doing the
actual data handling. The later can easily be tested in unit
tests.